Changed archive command to have subcommands create,list,extract
