+++ cmd/all/all.go
@@ -66,6 +66,4 @@ import (
 	_ "github.com/rclone/rclone/cmd/tree"
 	_ "github.com/rclone/rclone/cmd/version"
 	_ "github.com/rclone/rclone/cmd/archive"
-	_ "github.com/rclone/rclone/cmd/listarchive"
-	_ "github.com/rclone/rclone/cmd/extract"
 )
