+++ cmd/archive/create/files.go
@@ -1,5 +1,5 @@
 
-package archive
+package create
 
 import (
